# task-7/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "Nix Education Trainee Task API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/restricted/comments": {
            "get": {
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "summary": "Retrieves comments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Comment"
                        }
                    }
                }
            }
        },
        "/restricted/comments/{id}": {
            "get": {
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "summary": "Retrieves comment based on given ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Post"
                        }
                    }
                }
            }
        },
        "/restricted/comments/{id}/delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "summary": "Delete comments based on comment ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Comment"
                        }
                    }
                }
            }
        },
        "/restricted/comments/{id}/edit": {
            "put": {
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "summary": "Edit comment for given Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "post Id",
                        "name": "post_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "comment Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "E-mail",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Body",
                        "name": "body",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Comment"
                        }
                    }
                }
            }
        },
        "/restricted/posts": {
            "get": {
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "summary": "Retrieves posts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Post"
                        }
                    }
                }
            }
        },
        "/restricted/posts/add": {
            "post": {
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "summary": "Add post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User Id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Post Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Post Body",
                        "name": "body",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Post"
                        }
                    }
                }
            }
        },
        "/restricted/posts/{id}": {
            "get": {
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "summary": "Retrieves posts based on given ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Post"
                        }
                    }
                }
            }
        },
        "/restricted/posts/{id}/comments": {
            "get": {
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "summary": "Retrieves comments based on post ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Comment"
                        }
                    }
                }
            }
        },
        "/restricted/posts/{id}/comments/add": {
            "post": {
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "summary": "Add comment for given post Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "post Id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "E-mail",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Body",
                        "name": "body",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Comment"
                        }
                    }
                }
            }
        },
        "/restricted/posts/{id}/delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "summary": "Delete post based on post ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Post"
                        }
                    }
                }
            }
        },
        "/restricted/posts/{id}/edit": {
            "put": {
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "summary": "Edit post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User Id",
                        "name": "userid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Post Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Post Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Post Body",
                        "name": "body",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Post"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "database.Comment": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "postId": {
                    "description": "gorm.Model",
                    "type": "integer"
                }
            }
        },
        "database.Post": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "userid": {
                    "description": "gorm.Model",
                    "type": "integer"
                }
            }
        }
    }
}